vaar-api/agentic-api/health-assistant/requirements.txt
# Web Framework
fastapi==0.109.0
mangum==0.17.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16

# Vector Store
faiss-cpu==1.7.4

# Document Processing
pypdf==4.0.1
unstructured==0.12.2
python-multipart==0.0.6

# AWS
boto3==1.34.26

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2
